+++ HomebrewFormula/eilmeldung.rb
@@ -1,11 +1,11 @@
 class Eilmeldung < Formula
   desc "a feature-rich TUI RSS reader based on the newsflash library"
   homepage "https://github.com/christo-auer/eilmeldung"
-  url "https://github.com/christo-auer/eilmeldung/archive/refs/tags/1.1.0.tar.gz"
-  sha256 "566aa7ec5477cd66ecf88a30f3faaeb84f873628dd11e5724f54a89677298b3c"
+  url "https://github.com/christo-auer/eilmeldung/archive/refs/tags/1.2.0.tar.gz"
+  sha256 "replaced by CI on release of 1.2.0"
   license "GPL-3.0"
   head "https://github.com/christo-auer/eilmeldung.git", branch: "main"
-  version "1.1.0"
+  version "1.2.0"
 
   depends_on "pkg-config" => :build
   depends_on "rust" => :build